test: split test_sessions module
