+++ test/test_sessions.py
@@ -3,6 +3,7 @@ from datetime import datetime
 
 import httpx
 import pytest
+import test_roles
 
 from keycloak_admin_aio import KeycloakAdmin
 from keycloak_admin_aio._lib.utils import cast_non_optional
@@ -16,46 +17,45 @@ async def get_all_admin_cli_sessions(keycloak_admin: KeycloakAdmin):
     ).user_sessions.get()
 
 
-@pytest.mark.asyncio
-async def test_session_remove_handle(keycloak_admin: KeycloakAdmin):
-    sessions = await get_all_admin_cli_sessions(keycloak_admin)
-    assert len(sessions) > 0
-    sessions.sort(key=lambda s: cast_non_optional(s.lastAccess))
+class TestById:
+    """Test keycloak_admin_aio.sessions.by_id"""
 
-    async def delete_session(session_id: str):
-        try:
-            await keycloak_admin.sessions.by_id(cast_non_optional(session_id)).delete()
-        except httpx.HTTPStatusError as ex:
-            if ex.response.status_code != httpx.codes.UNAUTHORIZED:
-                raise ex
-            # Just terminated its own session
-
-    await asyncio.gather(
-        *[delete_session(cast_non_optional(session.id)) for session in sessions]
+    @pytest.mark.asyncio
+    @pytest.mark.dependency(
+        depends=[
+            test_roles.TestByNameLifeCycle.dependency_name("get", scope="session"),
+            "test/test_clients.py::test_get",
+            "test/test_clients.py::test_get_user_sessions",
+        ],
+        scope="session",
     )
-    keycloak_admin.access_token_expire = datetime.now()
-    try:
-        # Try doing anything
-        await keycloak_admin.roles.by_name("create-realm").get()
-    except httpx.HTTPStatusError as ex:
-        if ex.response.status_code == 401:
-            pytest.fail(f"Session removal not handled correctly: {ex}")
+    async def test_delete(self, keycloak_admin: KeycloakAdmin):
+        """Test keycloak_admin_aio.sessions.by_id.delete
 
+        Session delete is tested by deleting the session of keycloak_admin_aio itself.
+        """
+        sessions = await get_all_admin_cli_sessions(keycloak_admin)
+        assert len(sessions) > 0
+        sessions.sort(key=lambda s: cast_non_optional(s.lastAccess))
 
-@pytest.mark.asyncio
-async def test_only_one_session(keycloak_admin: KeycloakAdmin):
-    sessions = await get_all_admin_cli_sessions(keycloak_admin)
-    sessions_count_before = len(sessions)
-    async with KeycloakAdmin.with_password(
-        server_url="http://localhost:8080", username="testing", password="testing"
-    ) as kc:
+        async def delete_session(session_id: str):
+            try:
+                await keycloak_admin.sessions.by_id(
+                    cast_non_optional(session_id)
+                ).delete()
+            except httpx.HTTPStatusError as ex:
+                if ex.response.status_code == httpx.codes.UNAUTHORIZED:
+                    # Just terminated its own session
+                    return
+                raise ex
+
         await asyncio.gather(
-            kc.roles.by_name("create-realm").get(),
-            kc.roles.by_name("admin").get(),
-            kc.roles.by_name("admin").get(),
+            *[delete_session(cast_non_optional(session.id)) for session in sessions]
         )
-    sessions = await get_all_admin_cli_sessions(keycloak_admin)
-    sessions_count_after = len(sessions)
-
-    if sessions_count_before + 1 != sessions_count_after:
-        pytest.fail("KeycloakAdmin created instantiated to many sessions.")
+        keycloak_admin.access_token_expire = datetime.now()
+        try:
+            # Try doing anything
+            await keycloak_admin.roles.by_name("create-realm").get()
+        except httpx.HTTPStatusError as ex:
+            if ex.response.status_code == httpx.codes.UNAUTHORIZED:
+                pytest.fail(f"Session removal not handled correctly: {ex}")
